// src/core/http/generated/models/createPaymentGatewayPayload.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { CreatePaymentGatewayPayloadConfigJson } from './createPaymentGatewayPayloadConfigJson.ts';
import type { CreatePaymentGatewayPayloadProvider } from './createPaymentGatewayPayloadProvider.ts';
import type { CreatePaymentGatewayPayloadStatus } from './createPaymentGatewayPayloadStatus.ts';

export interface CreatePaymentGatewayPayload {
  name: string;
  provider: CreatePaymentGatewayPayloadProvider;
  status: CreatePaymentGatewayPayloadStatus;
  is_sandbox: boolean;
  public_key?: string;
  access_token?: string;
  client_id?: string;
  client_secret?: string;
  webhook_secret?: string;
  config_json?: CreatePaymentGatewayPayloadConfigJson;
}