// src/core/http/generated/models/updatePaymentGatewayPayload.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { UpdatePaymentGatewayPayloadConfigJson } from './updatePaymentGatewayPayloadConfigJson.ts';
import type { UpdatePaymentGatewayPayloadProvider } from './updatePaymentGatewayPayloadProvider.ts';
import type { UpdatePaymentGatewayPayloadStatus } from './updatePaymentGatewayPayloadStatus.ts';

export interface UpdatePaymentGatewayPayload {
  name?: string;
  provider?: UpdatePaymentGatewayPayloadProvider;
  status?: UpdatePaymentGatewayPayloadStatus;
  is_sandbox?: boolean;
  public_key?: string;
  access_token?: string;
  client_id?: string;
  client_secret?: string;
  webhook_secret?: string;
  config_json?: UpdatePaymentGatewayPayloadConfigJson;
}